suite/components/downloads/tests/chrome/test_space_key_retries.xul

<?xml version="1.0"?>
<!--
/* ***** BEGIN LICENSE BLOCK *****
 * Version: MPL 1.1/GPL 2.0/LGPL 2.1
 *
 * The contents of this file are subject to the Mozilla Public License Version
 * 1.1 (the "License"); you may not use this file except in compliance with
 * the License. You may obtain a copy of the License at
 * http://www.mozilla.org/MPL/
 *
 * Software distributed under the License is distributed on an "AS IS" basis,
 * WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License
 * for the specific language governing rights and limitations under the
 * License.
 *
 * The Original Code is mozilla.org code.
 *
 * The Initial Developer of the Original Code is
 *   the Mozilla Foundation.
 * Portions created by the Initial Developer are Copyright (C) 2008
 * the Initial Developer. All Rights Reserved.
 *
 * Contributor(s):
 *   Jens Hatlak <jh@junetz.de> (Original Author)
 *
 * Alternatively, the contents of this file may be used under the terms of
 * either the GNU General Public License Version 2 or later (the "GPL"), or
 * the GNU Lesser General Public License Version 2.1 or later (the "LGPL"),
 * in which case the provisions of the GPL or the LGPL are applicable instead
 * of those above. If you wish to allow use of your version of this file only
 * under the terms of either the GPL or the LGPL, and not to allow others to
 * use your version of this file under the terms of the MPL, indicate your
 * decision by deleting the provisions above and replace them with the notice
 * and other provisions required by the GPL or the LGPL. If you do not delete
 * the provisions above, a recipient may use your version of this file under
 * the terms of any one of the MPL, the GPL or the LGPL.
 *
 * ***** END LICENSE BLOCK ***** */

/**
 * This tests that the space key will retry a download in the UI.
 * This test was added in bug 474622.
 */
-->

<window title="Download Manager Test"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="test();">

  <script src="chrome://mochikit/content/MochiKit/packed.js"/>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"/>

  <script>
  <![CDATA[

function dlObs(aWin)
{
  this.mWin = aWin;
  this.wasCanceled = false;
  this.wasFinished = false;
}
dlObs.prototype = {
  observe: function(aSubject, aTopic, aData)
  {
    if ("timer-callback" == aTopic) {
      if (this.wasFinished) {
        // We're done!
        this.mWin.close();
        SimpleTest.finish();
      } else {
        // dispatch a space keypress to retry the download
        synthesizeKey(" ", {}, this.mWin);
      }
    }
  },

  onDownloadStateChange: function(aState, aDownload)
  {
    if (aDownload.state == Ci.nsIDownloadManager.DOWNLOAD_DOWNLOADING &&
        !this.wasCanceled) {
      this.wasCanceled = true;
      this.mWin.cancelDownload(aDownload);
    }

    if (aDownload.state == Ci.nsIDownloadManager.DOWNLOAD_CANCELED) {
      // We have to do this on a timer so other JS stuff that handles the UI
      // can actually catch up to us...
      var timer = Cc["@mozilla.org/timer;1"]
                    .createInstance(Ci.nsITimer);
      timer.init(this, 0, Ci.nsITimer.TYPE_ONE_SHOT);
    }

    if (aDownload.state == Ci.nsIDownloadManager.DOWNLOAD_FINISHED) {
      ok(this.wasCanceled,
         "The download was canceled, retried and then ran to completion");
      this.wasFinished = true;
      aDownload.targetFile.remove(false);

      var dm = Cc["@mozilla.org/download-manager;1"]
                 .getService(Ci.nsIDownloadManager);
      dm.removeListener(this);

      // We have to do this on a timer so other JS stuff that handles the UI
      // can actually catch up to us...
      var timer = Cc["@mozilla.org/timer;1"]
                    .createInstance(Ci.nsITimer);
      timer.init(this, 0, Ci.nsITimer.TYPE_ONE_SHOT);
    }
  },
  onStateChange: function(a, b, c, d, e) { },
  onProgressChange: function(a, b, c, d, e, f, g) { },
  onSecurityChange: function(a, b, c, d) { }
};
function test()
{
  var dm = Cc["@mozilla.org/download-manager;1"]
             .getService(Ci.nsIDownloadManager);

  function addDownload() {
    function createURI(aObj) {
      return (aObj instanceof Ci.nsIFile) ? Services.io.newFileURI(aObj) :
                                            Services.io.newURI(aObj);
    }

    const nsIWBP = Ci.nsIWebBrowserPersist;
    var persist = Cc["@mozilla.org/embedding/browser/nsWebBrowserPersist;1"]
                    .createInstance(nsIWBP);
    persist.persistFlags = nsIWBP.PERSIST_FLAGS_REPLACE_EXISTING_FILES |
                           nsIWBP.PERSIST_FLAGS_BYPASS_CACHE |
                           nsIWBP.PERSIST_FLAGS_AUTODETECT_APPLY_CONVERSION;

    var destFile = Services.dirsvc.get("ProfD", Ci.nsIFile);
    destFile.append("download.result");
    if (destFile.exists())
      destFile.remove(false);

    var dl = dm.addDownload(Ci.nsIDownloadManager.DOWNLOAD_TYPE_DOWNLOAD,
                            createURI("http://example.com/httpd.js"),
                            createURI(destFile), null, null,
                            Math.round(Date.now() * 1000), null, persist, false);

    persist.progressListener = dl.QueryInterface(Ci.nsIWebProgressListener);
    persist.saveURI(dl.source, null, null, 0, null, null, dl.targetFile, null);

    return dl;
  }

  // First, we clear out the database
  dm.DBConnection.executeSimpleSQL("DELETE FROM moz_downloads");

  // See if the DM is already open, and if it is, close it!
  var win = Services.wm.getMostRecentWindow("Download:Manager");
  if (win)
    win.close();

  const DLMGR_UI_DONE = "download-manager-ui-done";

  var testObs = {
    observe: function(aSubject, aTopic, aData)
    {
      if (aTopic != DLMGR_UI_DONE)
        return;

      SimpleTest.waitForFocus(function () { continueTest(aSubject) }, aSubject);
    }
  };

  function continueTest(win) {
      var doc = win.document;
      dm.addListener(new dlObs(win));

      addDownload();
      // we need to focus the download as well
      doc.getElementById("downloadTree").view.selection.select(0);
      Services.obs.removeObserver(testObs, DLMGR_UI_DONE);
  }

  // Register with the observer service
  Services.obs.addObserver(testObs, DLMGR_UI_DONE);

  // Show the Download Manager UI
  Cc["@mozilla.org/download-manager-ui;1"]
    .getService(Ci.nsISuiteDownloadManagerUI)
    .showManager();

  SimpleTest.waitForExplicitFinish();
}

  ]]>
  </script>

  <body xmlns="http://www.w3.org/1999/xhtml">
    <p id="display"></p>
    <div id="content" style="display:none;"></div>
    <pre id="test"></pre>
  </body>
</window>
